src/requirements.txt
streamlit
langchain-groq
python-dotenv
opencv-python
Pillow
pdf2image
pytesseract